Split executor evaluation-budget tests under the file-size ratchet.
Static CI failed because executor.node.test.ts grew past 2000 lines.
Move the four-slot budget saturation and queueable ALS-gap cases into
their own suite so the ratchet stays green without grandfathering.
